Added fixes for fill order test

+++ contracts/DepositDEX.sol
@@ -165,7 +165,7 @@ contract DepositDEX is IDepositDEX, UUPSUpgradeable {
   }
 
   function getBalance(address account, address collateral, uint112 price) public view returns (int112 balance) {
-    balance = (_balances[account][collateral] * int112(price)) / _INT_PRECISION;
+    balance = int112((int256(_balances[account][collateral]) * int112(price)) / _INT_PRECISION);
   }
 
   function getTotalBalance(address account, CollateralPriceData[] memory prices) public view returns (int112 balance) {
